feat(integrate): move integrate module into dynamics module
