+++ src/galax/dynamics/_dynamics/__init__.py
@@ -1,10 +1,10 @@
 """galax: Galactic Dynamix in Jax."""
 
-from . import mockstream, orbit
+from . import integrate, mockstream, orbit
 from .mockstream import *
 from .orbit import *
 
-__all__: list[str] = ["mockstream"]
+__all__ = ["integrate", "mockstream"]
 __all__ += orbit.__all__
 __all__ += mockstream.__all__
