+++ src/galax/dynamics/_dynamics/orbit.py
@@ -13,11 +13,12 @@ from astropy.units import Quantity
 
 from galax.coordinates import AbstractPhaseSpacePosition
 from galax.coordinates._utils import getitem_vectime_index
-from galax.integrate import DiffraxIntegrator, Integrator
 from galax.potential._potential.base import AbstractPotentialBase
 from galax.typing import BatchFloatScalar, BatchVec6, BroadBatchVec3, VecTime
 from galax.utils._shape import batched_shape
 from galax.utils.dataclasses import converter_float_array
+
+from .integrate import DiffraxIntegrator, Integrator
 
 if TYPE_CHECKING:
     from typing import Self
