+++ tests/smoke/dynamics/test_package.py
@@ -6,4 +6,9 @@ from galax.dynamics._dynamics import mockstream, orbit
 
 def test_all() -> None:
     """Test the `galax.potential` API."""
-    assert set(gd.__all__) == {"mockstream", *orbit.__all__, *mockstream.__all__}
+    assert set(gd.__all__) == {
+        "integrate",
+        "mockstream",
+        *orbit.__all__,
+        *mockstream.__all__,
+    }
